Load and store metamodel in .cso files
This huuge patch adds a new custom section to .cso files which stores a
binary-serialized metamodel, making us now a totally independent back
end. We only support function declarations without type parameters at
the moment, but the data is in the module proper.

We have to shell out to objcopy to grab the data, and then it's just
serialization/deserialization of byte streams.

Signed-off-by: Casey Dahlin <sadmac@google.com>