Unreachable else for SwitchCaseElseExpression
A SwitchCaseElseExpression has to be exhaustive, so if it has no else clause,
it must be guaranteed that one of the other clauses fires. We have to create
the else branch and mark it unreachable otherwise weird things happen when we
try to put together phi nodes to recover the value from the expression.

Signed-off-by: Casey Dahlin <sadmac@google.com>